部分修改
include\feature_map.h
include\feat\lab_feature_map.h
src\feat\lab_feature_map.cpp
金字塔修改

+++ FaceDetection/include/util/image_pyramid.h
@@ -141,7 +141,7 @@ class ImagePyramid {
     return img;
   }
 
-  const seeta::ImageData* GetNextScaleImage(float* scale_factor = nullptr);
+  const seeta::ImageData* GetNextScaleImage(int32_t* scale_factor = nullptr);
 
  private:
   void UpdateBufScaled();
